+++ maps/views.py
@@ -58,13 +58,16 @@ def delete_blog(request):
     t2 = threading.Thread(target=user_check)
     t2.start()
     if (request.method=='POST'):
-        unique_id = request.POST['hidden_unique_id']
-        mydb = mysqldb()
-        mycursor = mydb.cursor()
-        query = "DELETE FROM maps_blog WHERE unique_id='"+unique_id+"'"
-        mycursor.execute(query)
-        mydb.commit()
-        return redirect('/maps/bloginfo')
+        if request.user.id == None:
+            return redirect('/login/')
+        else:
+            unique_id = request.POST['hidden_unique_id']
+            mydb = mysqldb()
+            mycursor = mydb.cursor()
+            query = "DELETE FROM maps_blog WHERE unique_id='"+unique_id+"'"
+            mycursor.execute(query)
+            mydb.commit()
+            return redirect('/maps/bloginfo')
 
 def report_blog(request):
     t1 = threading.Thread(target=past_or_present)
@@ -72,54 +75,57 @@ def report_blog(request):
     t2 = threading.Thread(target=user_check)
     t2.start()
     if (request.method=='POST'):
-        unique_id = request.POST['hidden_unique_id']
-        print('unique_id:', unique_id)
-        mydb = mysqldb()
-        flag = 3
-        mycursor = mydb.cursor()
-        query = "select * from home_users where user in (select user from maps_blog where unique_id="+'"'+unique_id+'")'
-        mycursor.execute(query)
-        result = mycursor.fetchall()
-
-        reports = result[0][1]
-        identifying_id = unique_id.split('-')
-        reporting_id = identifying_id[4]
-
-        query = "select reported_map from home_users where user='"+str(request.user.username)+"'"
-        mycursor.execute(query)
-        check = mycursor.fetchall()
-
-        print(check[0][0])
-        checker = check[0][0].split()
-
-        if reporting_id in checker:
-            flag = 1
+        if request.user.id == None:
+            return redirect('/login/')
         else:
-            query = "select email from auth_user where username ='" + result[0][0] + "'"
+            unique_id = request.POST['hidden_unique_id']
+            print('unique_id:', unique_id)
+            mydb = mysqldb()
+            flag = 3
+            mycursor = mydb.cursor()
+            query = "select * from home_users where user in (select user from maps_blog where unique_id="+'"'+unique_id+'")'
             mycursor.execute(query)
-            email_list = mycursor.fetchall()
-            email = email_list[0][0]
-            # mail_seder(receiver_email=email, user=result[0][0], event='', date='', place='', flag=3)
-            t2 = threading.Thread(target=mail_seder, args=(email, result[0][0], '', '', '', 3))
-            t2.start()
-            if check[0][0] == '':
-                new_reports = str(int(reports) + 1)
-                query = "update home_users set report='" + new_reports + "' where user='" + result[0][0] + "'"
-                mycursor.execute(query)
-                mydb.commit()
-                query = "update home_users set reported_map='"+reporting_id+"' where user='"+str(request.user.username)+"'"
-                mycursor.execute(query)
-                mydb.commit()
+            result = mycursor.fetchall()
+
+            reports = result[0][1]
+            identifying_id = unique_id.split('-')
+            reporting_id = identifying_id[4]
+
+            query = "select reported_map from home_users where user='"+str(request.user.username)+"'"
+            mycursor.execute(query)
+            check = mycursor.fetchall()
+
+            print(check[0][0])
+            checker = check[0][0].split()
+
+            if reporting_id in checker:
+                flag = 1
             else:
-                new_reports = str(int(reports) + 1)
-                query = "update home_users set report='" + new_reports + "' where user='" + result[0][0] + "'"
+                query = "select email from auth_user where username ='" + result[0][0] + "'"
                 mycursor.execute(query)
-                mydb.commit()
-                new_reporting_id = check[0][0] + ' ' + reporting_id
-                query = "update home_users set reported_map='" + new_reporting_id + "' where user='"+str(request.user.username)+"'"
-                mycursor.execute(query)
-                mydb.commit()
-        Blogs = Blog.objects.all()
+                email_list = mycursor.fetchall()
+                email = email_list[0][0]
+                # mail_seder(receiver_email=email, user=result[0][0], event='', date='', place='', flag=3)
+                t2 = threading.Thread(target=mail_seder, args=(email, result[0][0], '', '', '', 3))
+                t2.start()
+                if check[0][0] == '':
+                    new_reports = str(int(reports) + 1)
+                    query = "update home_users set report='" + new_reports + "' where user='" + result[0][0] + "'"
+                    mycursor.execute(query)
+                    mydb.commit()
+                    query = "update home_users set reported_map='"+reporting_id+"' where user='"+str(request.user.username)+"'"
+                    mycursor.execute(query)
+                    mydb.commit()
+                else:
+                    new_reports = str(int(reports) + 1)
+                    query = "update home_users set report='" + new_reports + "' where user='" + result[0][0] + "'"
+                    mycursor.execute(query)
+                    mydb.commit()
+                    new_reporting_id = check[0][0] + ' ' + reporting_id
+                    query = "update home_users set reported_map='" + new_reporting_id + "' where user='"+str(request.user.username)+"'"
+                    mycursor.execute(query)
+                    mydb.commit()
+            Blogs = Blog.objects.all()
     return render(request, 'maps_blog.html', {'flag': flag, 'blog_data': Blogs})
 
 def display_blog_info(request):
